+++ src/visual/InteractiveCanvas.h
@@ -17,15 +17,15 @@ public:
     long long getFrequencyAt(float x);
     long long getFrequencyAt(float x, long long iqCenterFreq, long long iqBandwidth);
     
-    virtual void setView(long long center_freq_in, int bandwidth_in);
+    virtual void setView(long long center_freq_in, long long bandwidth_in);
     virtual void disableView();
     bool getViewState();
 
     void setCenterFrequency(long long center_freq_in);
     long long getCenterFrequency();
 
-    void setBandwidth(unsigned int bandwidth_in);
-    unsigned int getBandwidth();
+    void setBandwidth(long long bandwidth_in);
+    long long getBandwidth();
 
     MouseTracker *getMouseTracker();
     bool isMouseInView();
@@ -59,8 +59,8 @@ protected:
     bool ctrlDown;
 
     long long centerFreq;
-    unsigned int bandwidth;
-    unsigned int lastBandwidth;
+    long long bandwidth;
+    long long lastBandwidth;
 
     bool isView;
 	std::string lastToolTip;
